Move checkPlugins task into its own module.
